requirements.txt
Flask==3.1.0
Flask-APScheduler==1.13.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
flasgger==0.9.7.1
Flask-JWT-Extended==4.6.0
Flask-Cors==5.0.0
termcolor==2.4.0
PyJWT==2.9.0
python-dotenv ==1.0.1
psycopg2-binary ==2.9.10
pytz ==2025.1
requests == 2.32.3